+++ bench/Foldl.hs
@@ -3,11 +3,16 @@
 module Main (main) where
 
 import Control.Foldl hiding (map)
+import qualified Control.Foldl.NonEmpty as Foldl1
 import Criterion.Main
 import qualified Data.List
 import Prelude hiding (length, sum)
 import qualified Prelude
 import qualified Data.Foldable as Foldable
+import Data.Functor.Contravariant (Contravariant(..))
+import Data.Profunctor (Profunctor(..))
+import Data.List.NonEmpty (NonEmpty(..))
+import qualified Data.List.NonEmpty as NonEmpty
 
 main :: IO ()
 main = defaultMain
@@ -50,8 +55,26 @@ main = defaultMain
                 nf sumAndLength_foldl
             ]
         ]
+  , env (return $ 1 :| [2..10000 :: Int]) $ \ns ->
+      bgroup "1 :| [2..10000 :: Int]"
+        [ bgroup "handles" $ map ($ ns)
+            [ bench "fold (handles (to succ) list)" .
+                nf (fold (handles (to succ) list))
+            , bench "foldM (handlesM (to succ) (generalize list))" .
+                nfIO . foldM (handlesM (to succ) (generalize list))
+            , bench "NonEmpty.map succ" .
+                nf (NonEmpty.map succ)
+            , bench "Foldl1.fold1 (Foldl1.handles (to succ) (Foldl1.fromFold list))" .
+                nf (Foldl1.fold1 (Foldl1.handles (to succ) (Foldl1.fromFold list)))
+            ]
+        ]
   ]
 
+
+-- local definition to avoid importing Control.Lens.Getter.to
+to :: (Profunctor p, Contravariant f) => (s -> a) -> p a (f a) -> p s (f s)
+to k = dimap k (contramap k)
+{-# INLINE to #-}
 
 sumAndLength :: Num a => [a] -> (a, Int)
 sumAndLength xs = (Prelude.sum xs, Prelude.length xs)