+++ CHANGELOG.md
@@ -1,6 +1,9 @@
+- Add [Fold1 utilities](): `purely`, `purely_`, `premap`, `handles`, `foldOver`, `folded1`
+- Add pattern synonym `Fold1_` that makes the initial, step and extraction functions explicit.
+
 1.4.16
 
-- Add [`Control.Foldl.postmapM`]
+- Add [`Control.Foldl.postmapM`](https://github.com/Gabriella439/foldl/pull/205)
 
 1.4.15